requirement.txt
dashscope
modelscope
openai
opencv-python
python-dotenv
sageattention
xfuser
diffusers
accelerate
deepspeed
huggingface-hub
decord
transformers
einops
fastapi
ftfy
imageio
imageio-ffmpeg
lmdb
numpy
omegaconf
packaging
pandas
peft
pillow
pydantic
pyyaml
regex
requests
safetensors
scipy
sentencepiece
tqdm
typing-extensions
uvicorn
wandb
easydict